refactor: 이벤트 필터링 클래스 및 메서드명 변경
