+++ backend/src/main/java/com/dajava/backend/domain/heatmap/service/SolutionEventManager.java
@@ -1,4 +1,4 @@
-package com.dajava.backend.domain.heatmap.filter;
+package com.dajava.backend.domain.heatmap.service;
 
 import java.util.List;
 
@@ -6,11 +6,11 @@ import com.dajava.backend.domain.event.es.entity.SolutionEventDocument;
 import com.dajava.backend.domain.heatmap.validation.ScreenWidthValidator;
 import com.dajava.backend.domain.heatmap.validation.UrlEqualityValidator;
 
-public class SolutionEventFilter {
+public class SolutionEventManager {
 	private static final UrlEqualityValidator urlEqualityValidator = new UrlEqualityValidator();
 	private static final ScreenWidthValidator screenWidthValidator = new ScreenWidthValidator();
 
-	public static List<SolutionEventDocument> filter(List<SolutionEventDocument> events, String targetUrl, int widthRange) {
+	public static List<SolutionEventDocument> getValidEvents(List<SolutionEventDocument> events, String targetUrl, int widthRange) {
 		return events.stream()
 			.filter(event -> urlEqualityValidator.isMatching(targetUrl, event.getPageUrl())
 				&& screenWidthValidator.normalizeToWidthRange(event.getBrowserWidth()) == widthRange)
